refactor: move useStoryTimer to hooks
